AboutMe
I'm from Bulgaria, living and working in - Sofia. 
I was born in Yambol and came in Sofia after graduating. About a year ago I've found my professional field and am eager to give my best. I'm working in an advertising agency as a digital consultant in the Public relations department. Some of my client include Kaufland, Mastercard, Avon, Discovery and Eurosport. 
For me is very, very important to constantly learn and develop as a "digital" professional and in the context of ever changing digital technology and communications it's even more important to develop new skills. That's why I'm here.